refactor(core): move context to release folder
